// src/auditorias/interfaces/auditorias.interface.ts
export enum Representates {
    ASF = 'ASF',
    ASE = 'ASE',
    SFPF = 'SFPF'
}

export enum Etapa {
    Inicio = 'Inicio',
    Planeación = 'Planeación',
    Ejecución = 'Ejecución',
    Seguimiento = 'Seguimiento',
    Informe = 'Informe',
    Conclusión = 'Conclusión'
}

export enum TipoAuditoria {
    Directas = 'Directa',
    Conjuntas = 'Conjunta',
    Evaluaciones = 'Evaluación',
    Reviciones = 'Revisión'
}